+++ airflow/CHANGELOG.md
@@ -1,5 +1,9 @@
 # CHANGELOG - Airflow
 
+## 1.1.0 / 2020-02-22
+
+* [Added] Add airflow logs. See [#5405](https://github.com/DataDog/integrations-core/pull/5405).
+
 ## 1.0.0 / 2020-01-06
 
 * [Added] Add Airflow integration. See [#5232](https://github.com/DataDog/integrations-core/pull/5232).