+++ kube_metrics_server/CHANGELOG.md
@@ -1,5 +1,9 @@
 # CHANGELOG - Kube Metrics Server
 
+## 1.1.1 / 2020-02-22
+
+* [Fixed] Fix metric validation. See [#5581](https://github.com/DataDog/integrations-core/pull/5581).
+
 ## 1.1.0 / 2020-01-13
 
 * [Added] Make OpenMetrics use the RequestsWrapper. See [#5414](https://github.com/DataDog/integrations-core/pull/5414).